Upgraded to Webdriver-backed Selenium, using Webdriver APIs directly for many cases.  There is still some more work to do, but this is closer to passing than the current tests in our current environment.

+++ com.zutubi.pulse.acceptance/src/test/com/zutubi/pulse/acceptance/components/pulse/project/BuildSummaryTable.java
@@ -25,7 +25,7 @@ public class BuildSummaryTable extends SummaryTable
     public List<BuildInfo> getBuilds()
     {
         List<BuildInfo> result = new LinkedList<BuildInfo>();
-        int count = getRowCount();
+        long count = getRowCount();
         for (int i = 0; i < count; i++)
         {
             result.add(new BuildInfo(getRow(i)));
